qa_agent/requirements.txt
fastapi
uvicorn
streamlit
python-multipart
langchain
langchain-community
langchain-google-genai
chromadb
beautifulsoup4
unstructured
requests
tiktoken
selenium
python-dotenv
markdown
